test(zones): moved test subscriptions in setupServer, created resolver subdirectory and moved zone resolvers there
